Expand rustc lints
Adds a number of additional lints as a `#![warn(...)]` directive,
including `trivial_numeric_casts`, which caught a number of
redundancies.

+++ src/uint/sqrt.rs
@@ -10,7 +10,7 @@ impl<const LIMBS: usize> UInt<LIMBS> {
     ///
     /// Callers can check if `self` is a square by squaring the result
     pub const fn sqrt(&self) -> Self {
-        let max_bits = ((self.bits() + 1) >> 1) as usize;
+        let max_bits = (self.bits() + 1) >> 1;
         let cap = Self::ONE.shl_vartime(max_bits);
         let mut guess = cap; // ≥ √(`self`)
         let mut xn = {
